[FIX] l10n_pe_reports: Fixed sales report for cancelled documents
The `l10n_pe_edi_cancel_cdr_number` field undergoes cleaning during
the cancellation process, making it inappropriate to rely on it as
a confirmation of invoice cancellation.

The recent update involves utilizing the EDI state to validate both
the cancellation status and the prior signature of an invoice.

Additionally, SUNAT requires all amounts to be set to 0 for canceled
invoices in *SIRE* reports.

Extra: Considering the Odoo option to show all entries, now the draft
entries are added on the preview, but ignored on the TXT file

closes odoo/enterprise#54281

X-original-commit: 088826d8cb442474b5c9eef96a0a13202fa0297d
Signed-off-by: Josse Colpaert <jco@odoo.com>