[FIX] l10n_mx_edi: missing ref object while scheduled action
An error occurs when the customer runs the scheduled action
`Automatic update of state on the SAT (for invoices)`. This is because
`self.env.ref' is used instead of `self.env` for calling reference see [1].

[1]-https://github.com/odoo/enterprise/blob/saas16.4/l10n_mx_edi/models/l10n_mx_edi_document.py#L865

Error: "AttributeError:
       'str' object has no attribute 'transaction'"

This commit fixes the above issue by calling reference with
`self.env.ref' instead of `self.env`.

sentry:4707826242

closes odoo/enterprise#52519

X-original-commit: 22d74a453c3e47c7a733d2499e1f7dcd1f080bbf
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>