[FIX] account_accountant: bank reco widget: focus amount
There is a problem in the "Manual Operations" of the bank reconciliation widget with the focus of the clicked fields on line change.

Consider the following situation in the bank reconciliation widget:
The "Manual Operations" tab is already visible,
a line is selected and we click on the amount field of a different line.
Then the line should change and the corresponding input of the field should be focussed in the "Manual Operations" tab.
(The problem may happen with other columns / fields too.)

But currently it can happen that the field is not properly focussed.
The issue is that the focusing happens before the "Manual Operations" tab is "patched".
After the "patching" the focus is lost.

This commit corrects this issue by executing the focussing after the "patching".

Reproduce
  1. Go to bank reconciliation widget (with the default company).
     The first bank statement line should be selected:
     It has amount "$ 96.67" and label: "R:9772938  10/07 AX 9415116318 T:5 BRT: 100.00 C/ croip" (or similar).
     (I think it should always fail if the "Taxes" field is visible on some of the lines on right)
  2  Ensure that nothing is selected in the "Match Existing Entries" tab.
  3. Click on some line on the right.
     The "Manual Operations" tab should be selected now.
  4. Ensure there is no focus on the "Amount" field in the "Manual Operations".
  5. Click on the debit field of some other line.
  6. For a short moment (before the line is changed) the "Amount" field is focused.
  7. The line is changed and the "Amount" field is not focused.

closes odoo/enterprise#54308

X-original-commit: ab8f304756aaba0bf3f5a2a06f35caf2537f8d2a
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>