[FIX] planning: avoid traceback for non hr-user when opening avatar card
Steps
=====
- Install planning
- Log in as a planning user with no hr access
- Open the planning app
- Click on the avatar of a resource with roles linked to it

Issue
=====
An access right error message appears indicating that the current user
can't access the default_role_id field of the resource.

Cause
=====
This field of the resource.resource model is protected to user with hr
access rights. In the avatar card component it was read, whatever the
rights of the current user were.

Fix
===
A check on the access rights of the current user is added in the
component to avoid to read that field if it is not allowed.

closes odoo/enterprise#52343

Related: odoo/odoo#145443
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>