[FIX] stock_barcode: adjust stock_barcode tests
odoo/odoo#143570 changes the default order on the `stock.move.line`
model.
Which means that some tests using this order as index references won't
be aligned anymore and needs some adjustments. In order to avoid
changing the meaning of the test, only had a `sorted()` before the
comparisons to simulate the old default order.

closes odoo/enterprise#53320

X-original-commit: 70b88b2be797ef6f1dffbefbd38010c4125d7b7f
Related: odoo/odoo#147330
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>