[FIX] stock_barcode: adjust stock_barcode tests
odoo/odoo#143570 changes the default order on the `stock.move.line`
model.
Which means that some tests using this order as index references won't
be aligned anymore and needs some adjustments. In order to avoid
changing the meaning of the test, only had a `sorted()` before the
comparisons to simulate the old default order.

closes odoo/enterprise#53384

X-original-commit: 233b753051b218352f42485eb4d781d83343c96d
Related: odoo/odoo#147507
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>