[FIX] whatsapp: be defensive when fetching partner based on number
Currently when calling '_find_or_create_from_number' with a number that
cannot be formatted, the whole method crash due to an exception being raised
by 'phone_format'.

Instead of crashing, we now catch the error, and early return if the number
cannot be formatted. If we cannot do anything with it, better avoid creating
a dummy partner, and the method returns an empty recordset.

Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: e9113a8eb38bb7bc25df0af7c8ecbc8cfde99968
Part-of: odoo/enterprise#52840