[FIX] whatsapp: do not crash when creating wa message with invalid numbers
When creating a whatsapp.message, it tries to opt-in or optout the given
number. However if for some reason the number cannot be parsed and
formatted, the call to '_add' or '_remove' raises and the whole message
creation is aborted.

We don't want to crash just for that. We therefore catch the error when
trying to update opt-in status, and log if anything went wrong.

Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: bddd8b9aeea88a667fad7507aa072e3a927ddeb0
Part-of: odoo/enterprise#52905