[FIX] web_studio: report editor: remove branding from new nodes
In the report editor, edit a report which has inherited views in order to have some branding.

Try to add a node with the odoo editor just after a branded node.

Before this commit, the new node took the place, if they shared the same tag, of the old one when saving.
Note that this fix doesn't fix the fact that adding a node at that place won't work anyway,
because from the IrUiView point of view, there is no reliable way to place that new node
in any of the views in the inheritance chain.

Example:

Main View:
```xml
<div>
  <p/>
  <p/>
</div>
```

Inheriting view
```xml
<xpath expr="/div/p[2]" position="after">
  <h1 />
</div>
```

Result
```xml
<div>
  <p/>
  <p/>   <!-- In edition add a node before this one -->
  <h2/>
</div>
```

Naive result after the edition:
```xml
<div>
  <p/>
  <p/> <!-- Added by edition -->
  <h2 /> <!-- Added via the inheriting view -->
  <p />
</div>
```

Expected result:
```xml
<div>
  <p/>
  <p/> <!-- Added by edition -->
  <p/>
  <h2/>
</div>
```

It goes on and on, wherever even a very smart and hairy code would put the new node
that spot has every chance to be the wrong one anyway.

After this commit, the new node is not saved, as it cannot in the current state of the report editor
but at least the original node is not removed, in reality, nothing happens. It is maybe confusing
for the user, but the guess is that the confusion is a little bit lessened.

opw-3495170
opw-3602609
opw-3601359

closes odoo/enterprise#52884

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>