[FIX] sale_subscription: enable pay button when only one payment method
Issue:
======
When you go into subscription preview with only 1 payment method
available, the pay button will be disabled and you can't enable it.

Steps to reproduce the isssue:
==============================
- Make sure only 1 payment method available (demo)
- Create a subscription and make in into `to_invoice` status
- Go to preview

Origin of the issue:
====================
There is an extra payment_form which was used in set payment method or ,
manage payment methods. In 17.0, we redirect the user when clicking on
the button so we don't need that form anymore.

opw-3626037

closes odoo/enterprise#52421

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>