[FIX] website_helpdesk_livechat: disabled the helpdesk command when no
teams are available

This commit's purpose is to disable the helpdesk ticket creation command
when there are no teams available for the current connected user.
Currently, using this command this way will create a new ticket without
any team attached, rendering it useless.

step to reproduce:
- install helpdesk application
- open helpdesk
- open confifuration menu, select helpdesk team
- select 'customer care'
- activate the livechat option and saves
- leave helpdesk
- change from the current active company to any other company that
doesn't have any helpdesk team with the livechat feature enabled
- open the discuss app and select the 'general' discussion
- type /helpdesk your_ticket
a new ticket will be created without any team linked to it.

source of the issue:
when a team has the livechat option enable, the command is added
to the livechat regardless of the connected user's access right to the
team.

solution:
Only add the command to the list of livechat command if the user has
access to at least one team with the livechat feature enable.

affected version : 15.0 - master
task - 3631682
https://www.odoo.com/web#id=3631682&menu_id=4720&cids=1&action=333&active_id=4105&model=project.task&view_type=form

closes odoo/enterprise#68063

X-original-commit: 5dbcfd5ca70885b50ef363c60a4bdaefb4daead2
Related: odoo/odoo#175945
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>