[FIX] planning: allow create new shift in create select dialog in gantt cell
Before this commit, due to a bad rebase the override of `getSelectCreateDialogProps`
in `sale_planning` module has been reintroduced and so that override disables
the create button in select create dialog displayed when we click on gantt cell.

This commit removes that override to allow the user to be able to create a new shift
when he is in the select create dialog.

closes odoo/enterprise#52443

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>