[FIX] l10n_xx_reports: Localization reports missing "upgrade"
xx = pk,pt,se

Problem
---------
In 15.1, the financial reports where upgraded to a new version where we
can directly indent children line inside parents. Some reports are
missing that upgrade.

Objective
---------
Those reports should be converted.

Solution
---------
Apply the new formatting of the report in the XML definition using the
children_ids and line_ids tags. Remove the parent_id and
financial_report_id that are now unnecessary.
Target is 16.0 because a refactoring of account reports occurred in 16.0.
The only version that will miss the upgrade is saas-15.2.

task-3508362

closes odoo/enterprise#52856

X-original-commit: 46424c27440a881b4a45f0cfbd028016ce9eb369
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>