[FIX] l10n_us_hr_payroll: Remove useless chain of characters in ssnid
In the US it is necessary to show the Social Security number (SSN) and hide the first 5 numbers (ex: XXXXX1234). When we print the Payslip of an employee with or without an SSN in their record (hr.employee), an extra number chain is added to the printed Payslip.

The mask XXX-XX should only be applied when there is a number and it should overwrite the 5 first characters, not be pasted before.

Task link: http://www.odoo.com/web#id=3630674&model=project.task
task-3630674

closes odoo/enterprise#55056

X-original-commit: 005e1795914996236738e4a35e7f475cb7848948
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>