[FIX] l10n_cl_edi: correctly find account.move given document number
Currently, when trying to find an `account.move` from a document number,
we get the padding used on the latest `account.move` and use it, along
with the document type (and the document number)  to reconstruct a
move's name. However, the latest `account.move` could end up being a
generic journal entry, making the whole process fail.

opw-3490343

closes odoo/enterprise#52515

X-original-commit: 2334a3280b3b6e05bb61cd71f7554bb20b5c21f1
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>