[FIX] web_enterprise: restore ARIA attributes on the home menu
When the home menu was modified to use the command palette on 2a518bc6,
some ARIA attributes were lost that are still required. Such attributes
are intended to indicate what is the focused app when no search is
performed, i.e. the command palette has not been opened yet.

closes odoo/enterprise#55576

X-original-commit: ddb2c4dce9024439a2795780ec1c8fd4fc67b403
Related: odoo/odoo#152175
Signed-off-by: Pierre Paridans (app) <app@odoo.com>