[FIX] documents: prevent unnecessary computation of the write method
Before this commit:
Inside any workspace, select a document, and drag & drop it again into the
same current workspace.

Issue:
The current condition(located in _onDrop) fails to check if the workspace where
the document is being dragged to, is the same as the workspace it's already present.

Effects:
- 'orm.write' is being executed unnecessarily.
- Traceback is observed exclusively in 17.0 and master when similar steps are
followed inside TRASH workspace. (will be solved after this PR)

Reason:
- The workspaces inside the search panel are listed as list-items (<li>).
Inside those items we have their display names enclosed within the <header>.
On selection of any workspace (including trash), an active class is added to the
<header>.
- When the _onDrop event gets called, the condition before this commit tries
to check the presence of an 'active' class on the <li> and not on the <header>
element present inside it.

After this commit:
The updated condition removes unnecessary computation and eliminates the
traceback as well.

Task-3627626

closes odoo/enterprise#58772

X-original-commit: 407d87e8d92f9d5d3c7d167d978661d3b00a70eb
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Shubham Thanki (shut) <shut@odoo.com>