[FIX] project_enterprise,industry_fsm: fix skills group expand for tasks
Before this commit, when the user goes to the gantt view of task
in Field Service, the user got an error because of a wrong usage of
the domain in the group expand of user_ids.

This commit fixes the issue and also make sure the group_expand of
user_ids first take into account the search on skills to add users
in the gantt view of tasks when the group by is `user_ids` field.

task-3349411

closes odoo/enterprise#52684

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>