[FIX] l10n_be_codabox: depends on field on other module
The compute function '_compute_l10n_be_codabox_fiduciary_vat', depends
on field 'account_representative_id' which is declared in the module
account_report.

There is no depends or transitive depends to this module, so it raises
an 500 error when the module is not installed.

This error has been introduced in https://github.com/odoo/enterprise/pull/49968

closes odoo/enterprise#52674

X-original-commit: 09f11d96634094d18b107a8015acfdea8a15e04a
Signed-off-by: Cedric Snauwaert <csn@odoo.com>