[FIX] account_accountant: branch reconciliation account domain
Currently, when the current company is a branch of some of company and
the user is in bank reconciliation's manual operation or auto reconciliation wizard,

The account_id field will not be able to access the account from the root company's
accounts because of how the domain was previously set on them.

This commit aims to fix that behavior and allow their account_id field to access
the root company's account too.

closes odoo/enterprise#52562

Task-id: 3549961
Related: odoo/odoo#145860
Signed-off-by: William André (wan) <wan@odoo.com>