[IMP] account_reports: Improve report filters UX
The aim of this commit is simplify all the account report filters.
For several filters, we don't want to recompute the report
directly after we select an option. For example, the journal
filter that allows users to select from which journal they want
to get report lines, before this commit users have to select
one by one and recompute the report everytime.
Of course is not efficient at all and it can takes a lot of
times for big database. After this commit, users can select
several journals before recomputing the whole report.
We take advantage of this commit for cleaning a bit the
JS code of report filters and to have something more consistent.

This commit changes the precision_rounding parameter when we
make a float_is_zero by comparing custom currency rate and
actual rate, the bug occurs because some currencies have a
decimal precision bigger than what we decide for the comparison.

This commit remove an option ("ALL") from a l10n_ar filter.
The idea is that we want to keep something consistent between all
the filters and it's the only filter that implement an option "ALL".
Now, when users don't select any filters or all the filter, it's
implied "ALL" without adding this new option.

task-3710483

closes odoo/enterprise#55012

Related: odoo/odoo#154697
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>