[ADD] l10n_mx_edi_stock{_extended}_30: Carta Porte 3.0
This commit adds the new mandatory fields according to Version 3.0 of the delivery guide.
It also extends the functionality of the delivery guide to Receipts.

task-3608164

Done afterwards by jco and checked by bib through testing by lao:
- the figuras need an extra name for which we took the name to the
partner
- importations need origen and destination ubicaciones switched
in the XML
- we moved the tests to l10n_mx_edi_stock_extended_30 because then
it does not depend on the modules installed if the test will pass or not
- Municipio is added as some PACs need it, but we removed the Colonia in the
Domicilio of the Ubicaciones.  The inheritance is a bit complicated as
we need incoming in the base module in which we switch incoming
and outgoing, but we add the municipio later.  For the incoming we added it
already in the base, but with a method, so it can return None if we do not
have the municipio already.
- The old customs number was not used, so we just create new fields with clear
purpose for the picking itself and use the new fields.  There was a position="replace"
too much.
- Everything is based on the old template which has a different namespace, but
we need to do the XSLT on the new namespace, so we needed to move the replace
- Needed to change type fields to goods_direction for style

closes odoo/enterprise#53506

X-original-commit: 6187af6d26477d88c49de8661157de328f25d0e1
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>