[FIX] web_studio: fix crash when editing rainbow effect on button
This commit fixes a crash occuring when editing the 'effect' attribute being set
on a button.

Other issues have been fixed by this commit, such as the current value of the fadeout,
that was not given properly to the component, and the onSelect props that was wrong,
causing another crash. The white border set by default on the component has also been
removed to match the sidebar style.

A test has also been added to verify that editing the effect from the sidebar is
sending the correct values to the 'edit_view' call.

task-3613476

closes odoo/enterprise#54110

X-original-commit: eeebbae121860476859668813314e6f0b0c4a6e1
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>