[REM] {website_}appointment{_crm,_hr_recruitment}: remove apt tag
Remove the "Appointment Portal" tag that was automatically added to all calendar events
related to an appointment type as this tag is now useless.
Those events can indeed be retrieved by filtering on the "appointment_type_id" field.
NB: This tag is called "Appointment" when website_appointment is installed.

Also prevent adding the "Interview" tag on the calendar events related to an
appointment type when it has an applicant as those can be retrieved using the
"applicant_id" field.

Task-3615058

closes odoo/enterprise#51668

Related: odoo/odoo#145908
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>