[FIX] l10n_ro_reports: Fix unbalanced Balance Sheets.
In Romania, there are two balance sheets: one for general companies
and one for companies applying IFRS. This commit aims to fix both.

  1. Balancing the general balance sheet
According to the Balance Sheet diagnosis tool in #36838, some accounts
were not referenced correctly. This is now fixed. We also re-do the
Retained Earnings / Current Year Earnings with cross-reports to the P&L.

  2. Balancing the IFRS balance sheet
In addition to the issues encountered with the general balance sheet,
the IFRS balance sheet also suffered from the problem that officially
it should be used with a CoA dedicated for IFRS, which does not contain
some of the accounts of the general CoA. Our solution is to nevertheless
add those accounts to the Balance Sheet in appropriate locations.

  3. Refactor the reports to use 1234D/1234C account code notation
When the Romanian reports were first written, we thought that certain
lines needed to contain a sum of debits or a sum of credits of AMLs of
certain accounts. However, it is now certain that what was intended was
the credit or debit balance of those accounts. We therefore remove the
domain_formula expressions and replace them with account_codes_formula
expressions.

  4. Adapt the l10n_ro_saft tests to the CoA changes in the community PR

Community PR: https://github.com/odoo/odoo/pull/145947

closes odoo/enterprise#52907

Taskid: 3634044
X-original-commit: bbde050ecc640e169d048200c8aa027ec3193be5
Related: odoo/odoo#146572
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>