[IMP] {timesheet,}_grid: Color code for the totals in all timesheets
Before this commit:
Currently, we have a color code for the daily total of timesheets encoded per employee,
but that doesn't apply to the total number for the period.

The daily color code can be misleading because the employee may be missing hours on one day,
but that is balanced out with overtime on another day and ultimately has encoded the required
number of hours according to their contract for the week/month.

The color code for the totals would help users identify at a glance employees who are missing
hours - especially when employees are not working standard 40h/week.

After this commit:
The following color for the totals were applied when grouping by employee:
- orange if overtime
- green if exact hours
- red if missing hours

task-3632025

closes odoo/enterprise#52608

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>