[FIX] *: remove escaped newlines from translated strings
The parser of the library we use to extract strings to be translated
(python-babel) doesn't handle escaped newlines in the same way as the
JavaScript parser. The resulting mismatch makes the content that uses
them untranslatable.

for example:
```js
const hey = _t("slt \
çv ?"
);
```
is exported as "slt \nçv ?", while the content of the variable is
"slt çv ?".

This commit removes faulty occurrences from the code base and re-exports
the corresponding POT files.

*: helpdesk

closes odoo/enterprise#54081

X-original-commit: 0bad1b0b640baac25b289e3301bfd6b6546e8ce0
Related: odoo/odoo#148931
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>