[FIX] account_reports: Journal Report Analytic Groupby bugged
Descritpion of the issue/feature this commit addresses:

At the moment, trying to use the analytic groupy on the Journal Report results
in a traceback. It is not possible to use it at all.

Steps to reproduce:

1-Activate the Analytic Accounting in the settings
2-Activate the Analytic Groupby in the Journal Report's options
3-Go to the Journal Report and apply any Analytic Groupby filter
4-A traceback shows up

Desired behavior after this commit is merged:

This commit handles the issue by adding a check at the very start of the custom
handler of the Journal Report. This new check verifies that no analytic groupby
filter is being applied on the report. If it is the case, an error is raised to
tell the user the feature is not supported by that report.

While at it, the Horizontal Grouping and Period Comparison options have also
been disabled because their rendering was making the report unreadable and they
added no value to it.

With this commit, the Journal Report can't use Analytic Groupby, Horizontal
Grouping or Period Comparison.

task-3600349

closes odoo/enterprise#54476

X-original-commit: dc16f0dd4aad01ab6b5a1a8e2ed116a64e9bb783
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>