[FIX] helpdesk: can delete stage in grouped kanban
Since the refactoring of the relational model [1], group.resModel
is the name of the model of the records, whereas before it was the
name of the relation of the field we were grouped by (when grouped
by a many2one field). Moreover, group.resId no longer exists, the
id of the group is now stored in group.value, by consistency when
grouped by other field types than many2one.

Before this commit, a crash occurred when the user tried to delete
a stage (a column) in the Helpdesk kanban view, when it was grouped
by state_id.

[1] https://github.com/odoo/odoo/pull/114024

OPW-3629005

closes odoo/enterprise#53330

X-original-commit: 10fa267e0b83e078e376b00c23556a76ea870a1b
Related: odoo/odoo#147355
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>