[IMP] l10n_mx_edi: duplicate check using fiscal folio
As every fiscal folio on a vendor bill in Mexico is unique, we can
easily check that to find duplicates.

This change extends the default duplicate vendor bill check to first
see if it can find duplicates with the same folio fiscal. If it doesn't
find any, it will apply the generic duplicate check.

task-3590442

Part-of: odoo/enterprise#53128