[IMP] l10n_mx_edi: duplicate check using fiscal folio
As every fiscal folio on a vendor bill in Mexico is unique, we can
easily check that to find duplicates.

This change extends the default duplicate vendor bill check to first
see if it can find duplicates with the same folio fiscal. If it doesn't
find any, it will apply the generic duplicate check.

task-3590442

X-original-commit: a493d15e202a59231ccb6684777454365a8cffee
Part-of: odoo/enterprise#56177