[FIX] account_accountant: analytic distrib exchange diff line
When having an analytic distribution model set with account
prefix matching exchenge diff accounts, this model is not taken
into account at the creation of the exchange diff line on the
reconciliation widget.
With this commit, we backport the commit d1b3b6959f5f5e72e97a18209334b2b6e464f1c1.
We also keep trace of the analytic distributions selected
on the line to avoid them to be erased at the creation of the
exchange diff move.

Steps:

- Have a distribution model for exchange diff account prefix
- Have a payment and a statement line in two different currencies
  and two different rate date
- In bank reco widget, select the statement line and select
  payment as counterpart
- Select the exchange diff line created
--> The analytic distribution set on the model is not set on the form,
- Now select another analytic distribution, different than the one
  set on the model
- Validate the reconciliation
- Go to the exchange diff move
--> The selected distribution has been replaced by the one set one the
    model

opw-3603668

closes odoo/enterprise#53350

X-original-commit: 7032831905d612cd68ba76eaa0928c2638ea67c7
Related: odoo/odoo#147404
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>