[IMP] l10n_ar_edi: check all the webserives available
Task Adhoc: 31784
Task latam: 1131

Description of the issue/feature this PR addresses:
It is necessary to make the connection test of all the webservices available in AFIP. In runbot Odoo 17.0 with module l10n_ar_edi installed and taking position in company "(AR) Responsable Inscripto" go to "Accounting / Configuration / Settings" and in section "Argentinean Localization" click on "Test Connections" (in "AFIP Web Services"). Then a pop up will show the status of the connection of the webservices wsfe, wsfex, wsbfe, wscdc but is needed to show the status of the connection of all webservices available.

Current behavior before PR:
It is shown the status of the connection of the webservices wsfe, wsfex, wsbfe, wscdc.

Desired behavior after PR is merged:
Show the status of the connection of all webservices available.

closes odoo/enterprise#52670

Signed-off-by: Josse Colpaert <jco@odoo.com>