[IMP] web_studio: improve app creator design
This commit improves the design of the app creator to match the latest
Odoo design (called Milk). We also use the SelectMenu component to display
the list of colors and icons available, to have a better UI and support
search and keyboard navigation.

The commit also adds the entire set of FontAwesome icons available for the
icon creator. The list is generated by searching into the known css rules to
match '.fa-' classnames, corresponding to icon classes. Some glyphs are used
for multiple classnames. The selection menu filter them to only show the icon
once, but keeping the text from being searchable.

E.g. fa-bank, fa-university and fa-institution use the same icon. The component
only list the fa-insitution with the 'Institution' tooltip. But searching 'bank'
will filter and show the icon in the list.

Additionnaly, because of the changes introduced in this commit, the 'Add a
button' dialog was outdated, since its height became very long. Instead, the
new component to select an icon is used.

Finally, some cleaning of dead code has been done to the IconCreator component
file. Tests have been adapted to match the latest layout, the updated colors,
and a new test have also been added to verify the filtering of icons using
the search feature.

task-3612833

closes odoo/enterprise#54400

X-original-commit: 0fa16a3bd678d8864fe66a72021d61fb3d04a092
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>