[FIX] l10n_cl_edi_pos: fix about customer and invoices
Prior to this commit, if you try to invoice a new contact created
in the PoS, but forgot to put his RUT,
then you get an error and it gets stuck there "thinking".

This commit adds the RUT number to the required fields in the chilean
localization of the PoS, so that it is not possible to invoice without
it.

This commit also adapts the partner_editor override in the chilean and
mexican localization to adapt it to the fix of the community branch.
This fix makes it impossible for a user to modify commercial fields of
a child partner.

closes odoo/enterprise#53061

X-original-commit: 9e2fccb0e457adbbb1abe0add04116c42770678e
Related: odoo/odoo#146807
Signed-off-by: Robin Heinz (rhe) <rhe@odoo.com>
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>