[FIX] l10n_at_reports: fix expressions for upgrade
In PR #47198 some field entries were removed from some expression
records in the XML.
This leads to problem with upgrade since only "positive changes"
(values that occur in the XML) are written to the DB.
Values that are absent are not set to False or to their default.
This was corrected in this commit.

A somewhat special case are the 'domain_formula' fields; they create
expressions. Setting their value to False explictly unlinks the
associated expressions.

Reproduce (leftover subformula causing traceback in P&L)
  1. Check out commit fc8a6c8344f95b20947188f341b16a648cdd1433
     (last commit before PR #47198)
  2. Start and install l10n_at_reports
  4. Check out commit before this commit
  5. Start with -u l10n_at_reports
  6. AT profit & loss
  7. Traceback

Reproduce (for domain_formula)
  1. Check out commit fc8a6c8344f95b20947188f341b16a648cdd1433
     (last commit before PR #47198)
  2. Start and install l10n_at_reports
  3. Check Accounting > Configuration > Accounting Reports
     > Bilanzdarstellung nach § 224 UGB (AT)
     > 4. sonstige Forderungen und Vermögensgegenstände
  4. Check out commit before this commit
  5. Start with -u l10n_at_reports
  6. Check again: Accounting > Configuration > Accounting Reports
     > Bilanzdarstellung nach § 224 UGB (AT)
     > 4. sonstige Forderungen und Vermögensgegenstände
     There are 2 expressions now but there should be just 1
     (the one from step 3 should be gone now)

closes odoo/enterprise#52844

X-original-commit: 685284524d62aa450e44e782e8b6f22c8ed61d3b
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>