[FIX] knowledge: keep the new embedded view title
This commit fixes a bug where an embedded view's title would no longer be saved
after being set to an empty value. Meaning that after removing the title it is
no longer possible to set as anything else than its default value.

This was caused by a condition stating that we could only update the
display_name inside the behavior props if its value already existed, which is
incorrect when a user removes it.
Now we change this value no matter what happens.

task-3637710

closes odoo/enterprise#53168

X-original-commit: acc84caa922f06baa0b113c0f30ead518ffc707c
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>