[FIX] account_online_synchronization: use a better async system
Cron_trigger on saas can take up to 15-20min to start as there is only
one thread responsible for waking up cron for thousands of databases.
Hence the async fetch we did in v17 is not really good in practice
(clicking on manual fetch and triggering a cron_trigger might take
20 mins and import 0 transactions).

Now we will, on manual import, fetch the first 100 transactions
(by date), process them and show them to the user. Then, a cron will
asyncronously check if other transactions have to be imported.

task-3613030

closes odoo/enterprise#52902

X-original-commit: 8475f92f4b6d977798c4f9d2485f7e1835c43c04
Signed-off-by: Cedric Snauwaert <csn@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>