[FIX] appointment: fix traceback grouping gantt by partner_ids
Currently if you group the gantt view on calendar events by partner_ids
you get a traceback because we try calling a method that is not named
properly.

As the intended feature is not supported  and we don't expect anybody
to have been using the method nor is it used anywhere internally
we remove both the python method that should have been called and
its caller in js.

It was a remnant of trying to make grouping by partner_ids the default
Which ended up being scrapped except for this part.

Although it is being reestablished in 17.0 and up with this task

task-3452277

closes odoo/enterprise#52829

X-original-commit: 227736b622dcce7c239c174bfa38ae56a5268cd0
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>