[FIX] account_reports: extra _get_domain param missing
When doing the fw-port of adding a 'Not Started' column [1] to version
17.0 [2], we added the possibility to audit that column as well.

However in doing so, adding a new parameter `filter_not_started` to the
`_get_domain` method, we forgot to adapt one method call to include that
paramater, causing clicking the caret action in the report to fail.

In order to fix the issue and simplify things, the parameters
`filter_already_generated` and `filter_not_started` get a default value
of `False` now, and the method calls are adapted accordingly.

_Bug reported by RIGR_

[1] 563404395c96650b871e293509af77acc811a7ff
[2] b3eaaa55b4550f1fefbd6dd472d9e8b8c4d21384

closes odoo/enterprise#52760

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>