[FIX] web_studio, *: keep dark mode setting after studio session
*: web_enterprise,website_enterprise

When opening Studio, the color scheme is reset to light mode.

Before this commit, if the UI was in dark mode before opening Studio,
when leaving it the dark mode isn't applied back.

This commit fixes it by keeping track separately of the configured
color scheme and the effective one in the "color_scheme" service.

Steps to reproduce:
* Switch to dark mode
* Open Studio
* Close it
=> color scheme is still "light"

opw-3419363

closes odoo/enterprise#53155

X-original-commit: 16780f997f26a3f137b6fc6a6890e07f9a62692c
Signed-off-by: Pierre Paridans (app) <app@odoo.com>