[FIX] sale_subscription: partially paid subscription saving token
Before this commit, when a payment link was used to partially pay a renewal subscription in order to confirm it, a token was set at the confirmation of the renewal. The consequence of this is that we receive a partial payment and then the remaining amount. We want to save the token only if the subscription is fully paid.

After this commit, when confirming a renewal subscription, the token is only saved if the sum of partial payments equal the total amount of the subscription. This way the invoice cron won't charge the total amount on the token.

closes odoo/enterprise#52857

Ticket-id: 3499685
X-original-commit: b00a10a19243bb842d7dd8fdeac06d85b792ed59
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>