[IMP] *: Stop using url fragment for state of the page
Before this commit, we used url fragment (#) to store the state of
the page in the form of pairs of key/values.

The aim of this commit, is to be more coherent with the web practices.
The fragment is traditionally used to indicate the id of the HTML
element (an anchor) on the page that the browser should scroll to, this
is automatically done by the browser.

This commit, change the current practice and start using the
query/search string of the url to store the state of the page.

Part of task-id : 3557575

closes odoo/enterprise#52662

Related: odoo/odoo#146070
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>