[FIX] whatsapp: allow WA admins to chose / change template model
Currently model choice on template model is done using directly the 'model_id'
field. However as it is a m2o towards 'ir.model' only sys admins have access
to it. As a WhatsApp admin you encounter en error when trying to change
anything related to the model.

We fix it by using the 'model' field instead. It is now a selection instead
of a char field, based on existing models. This should be update-safe as
it was a related on model before this change, so its value should not change.

'model_id' field is now computed based on 'model', inversing the behavior.
Override of create / write is added in order to keep working with direct
'model_id' update and be backward compatible.

Task-3640281
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: 74105d3fc1448cbd1b0942fd88759cf87f96bb7c
Part-of: odoo/enterprise#52905