[FIX] web_studio: add new digits type in supportedOptions
Before this commit, the digits options (eg: in FloatField) was of type
string. This means that when editing the value with studio, the user
could type anything. The FloatField would then fail when being
instanciated after leaving studio when editing a record.

When passing the digits option, it should only contain an array of two ints.

The first number is supposed to represent the total number of digits and
the second one represent the decimal precision.

The total number isn't actually used in the JS codebase.

Solution
========

- Introduce a new option type: digits
- Only allow integers between 0 and 20 as that value is passed to
  `toFixed`
- If the entered value is invalid, set a red border on the input and do nothing

opw-3614919

closes odoo/enterprise#53865

X-original-commit: 160cf85a216284ae783c0acc57928f78e4b08e1a
Related: odoo/odoo#148571
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>