[IMP] l10n_fr_reports: remove 'fetch edi files' button
We need to get rid of this button because it could raise potential legal
issues (see linked task).

It is marked invisible s.t. new databases won't have it. For the
existing databases, we rewrite the returned action to return the content
of the first xml attachment on the async export.

Also remove the IAP controller `_get_recipient_reports_v2` to make sure
it can no longer be called.

task-3640693

closes odoo/enterprise#54750

X-original-commit: b2cbd02693f5d1d539e6ce99a09cf3560b3bc152
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>