[IMP] hr_contract_salary: allow feedback from salary configurator
An offer is often refused. Currently, they need to be tracked somewhere else,
and managed manually. It should be made easier, so that the user can give
feedback about the offer to the hr manager. The HR manager can then refuse the
offer on the backend and select a refusal reason. Having defined refuse reasons
will allow to make dashboards.

task-3619658

closes odoo/enterprise#52806

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>