[FIX] knowledge: fix traceback when restoring an article
When the user clicks on the option "Restore from Trash" option from the
"more options" panel of the Knowledge editor, the system calls the
`action_unarchive_article` action and execute the action returned.
Currently the action `action_unarchive_article` does not return any
action, causing the action service to receive `undefined` and subsequently
trigger an error.

This commit will address this issue by reloading the view instead of
executing the action returned by `action_unarchive_article`.

Steps to reproduce the issue:
1. Open an article in the editor
2. Click on "Delete" button from the "more options" menu
3. Click on "Open the Trash"
4. Click on the article you deleted
5. Click on "Restore from Trash" from the "more options" menu
=> TypeError: `action` is undefined

TO BE: The editor is expected to reload the record and update the view.
Additionally, the banner warning users about the imminent deletion of
the article should be removed upon restoration.

task-3619503

closes odoo/enterprise#52814

X-original-commit: e3ef920872fdba9bc730465a46ff49cb099e1289
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>