[IMP] whatsapp: allow to add sidebar categories dynamically
Before this PR, a registry was used to list the categories shown in the
discuss sidebar. This can be done easily with mail records instead.
Doing so will allow to add new categories dynamically. Moreover, this is more consistent with the
discuss code base.

part of task-3640730

closes odoo/enterprise#52873

Related: odoo/odoo#146451
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>